requirements/requirements.txt
pre-commit
black
flake8

numpy
numba
opencv-python
cached-property
python-telegram-bot
